ft: factor out common code
 - factor out common code

[Delivers #170843732]
